requirements.txt
# Common / shared
python-dotenv==1.1.1
pydantic==2.11.0
requests==2.31.0
pandas==2.1.1
numpy==1.26.0

# API side
fastapi==0.112.0
uvicorn==0.23.2
SQLAlchemy==2.0.41
alembic==1.11.2
psycopg2-binary==2.9.9
passlib[bcrypt]==1.7.4
PyJWT==2.9.0
pydantic-settings>=2.11.0,<3.0
email-validator==2.0.0

# UI side (Streamlit, etc.)
streamlit==1.29.0
plotly==5.18.0
httpx==0.24.1
